requirements.txt
bcrypt
psutil
mysql-connector-python
docker
Flask
Werkzeug
python-dotenv